fix some build warnings

+++ modules/audio/builtin/cpp/audioplayer_playblockingBuiltin.cpp
@@ -75,13 +75,13 @@ Nelson::AudioGateway::audioplayer_playblockingBuiltin(
             if (start > objPlayer->getTotalSamples()) {
                 Error(_W("Invalid range."));
             }
-            objPlayer->play(start);
+            objPlayer->play((int)start);
         } else {
             if (start > objPlayer->getTotalSamples() || start > end
                 || end > objPlayer->getTotalSamples()) {
                 Error(_W("Invalid range."));
             }
-            objPlayer->play(start, end);
+            objPlayer->play((int)start, (int)end);
         }
     }
     do {
